Add USES=node[:args] for specifying a dependency on Node.js
This is based lightly on Mk/Uses/python.mk. Add build, run, test
dependencies based on the arguments received. A port can specify which
major version it needs. If the port does not explicitly set a version,
use a default which can be overridden by the user. Current possible
values are 4, 6, 8, and 9, which correspond to the active LTS and
Current releases from upstream.